- Updated CodeLite project - Some fixes for compilation with gcc
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@919 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/Compression.cpp
@@ -249,7 +249,7 @@ bool Compression::BZip2Compress(MemChunk &in, MemChunk &out) {
 	out.clear();
 
 	// Allocate a work buffer big enough to guarantee success
-	size_t bufferlen = in.getSize() + (in.getSize()>>6) + 1024;
+	unsigned bufferlen = in.getSize() + (in.getSize()>>6) + 1024;
 	char * buffer = new char[bufferlen];
 	int ok = -1;
 	if (buffer) {
